Renamed signatureParser.ts to signature_parser.ts
